split java/bootstrap generator
