+++ generators/java/generators/bootstrap/templates/src/main/java/_package_/GeneratedByJHipster.java.ejs
