+++ generators/server/generator.js
@@ -41,7 +41,6 @@ import {
   GENERATOR_DOCKER,
   GENERATOR_GATLING,
   GENERATOR_GRADLE,
-  GENERATOR_JAVA,
   GENERATOR_LANGUAGES,
   GENERATOR_MAVEN,
   GENERATOR_SERVER,
@@ -136,7 +135,6 @@ export default class JHipsterServerGenerator extends BaseApplicationGenerator {
     if (!this.delegateToBlueprint) {
       await this.dependsOnJHipster(GENERATOR_BOOTSTRAP_APPLICATION);
       await this.dependsOnJHipster(GENERATOR_COMMON);
-      await this.dependsOnJHipster(GENERATOR_JAVA);
     }
   }
