+++ generators/spring-boot/generator.ts
@@ -77,6 +77,7 @@ export default class SpringBootGenerator extends BaseApplicationGenerator {
 
     if (!this.delegateToBlueprint) {
       await this.dependsOnJHipster(GENERATOR_SERVER);
+      await this.dependsOnJHipster('jhipster:java');
     }
   }
